+++ src/frontend/packages/lib_components/src/common/queries/fetchList.tsx
@@ -3,6 +3,7 @@ import { QueryFunctionContext } from 'react-query';
 
 import { useJwt } from 'hooks/stores/useJwt';
 
+import { fetchResponseHandler } from './fetchResponseHandler';
 import { fetchWrapper } from './fetchWrapper';
 
 export type FetchListQueryKey =
@@ -38,9 +39,7 @@ export const fetchList = async <T,>({
     },
   );
 
-  if (!response.ok) {
-    throw new Error(`Failed to get list of ${name}.`);
-  }
-
-  return (await response.json()) as T;
+  return await fetchResponseHandler(response, {
+    errorMessage: `Failed to get list of ${name}.`,
+  });
 };