+++ src/frontend/packages/lib_components/src/common/queries/fetchOne.tsx
@@ -2,6 +2,7 @@ import { QueryKey, QueryFunctionContext } from 'react-query';
 
 import { useJwt } from 'hooks/stores/useJwt';
 
+import { fetchResponseHandler } from './fetchResponseHandler';
 import { fetchWrapper } from './fetchWrapper';
 
 /**
@@ -23,9 +24,7 @@ export const fetchOne = async <T,>({
     },
   });
 
-  if (!response.ok) {
-    throw new Error(`Failed to get ${endpoint}.`);
-  }
-
-  return (await response.json()) as T;
+  return await fetchResponseHandler(response, {
+    errorMessage: `Failed to get ${endpoint}.`,
+  });
 };