+++ src/frontend/packages/lib_components/src/common/queries/index.ts
@@ -4,5 +4,6 @@ export * from './deleteOne';
 export * from './fetchList';
 export * from './fetchOne';
 export * from './fetchWrapper';
+export * from './fetchResponseHandler';
 export * from './metadata';
 export * from './updateOne';