+++ src/frontend/packages/lib_components/src/common/queries/updateOne.tsx
@@ -1,5 +1,6 @@
 import { useJwt } from 'hooks/stores/useJwt';
 
+import { fetchResponseHandler } from './fetchResponseHandler';
 import { fetchWrapper } from './fetchWrapper';
 
 interface Variables<K> {
@@ -23,13 +24,5 @@ export const updateOne = async <T, K>({
     body: JSON.stringify(object),
   });
 
-  if (!response.ok) {
-    if (response.status === 400) {
-      throw { code: 'invalid', ...(await response.json()) };
-    }
-
-    throw { code: 'exception' };
-  }
-
-  return (await response.json()) as T;
+  return await fetchResponseHandler(response);
 };